LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/shape-outside/shape-image/shape-image-028-expected.txt

Verify that an image valued shape-outside layout is updated after the image has been loaded. This test checks that the left edge of the "Hello World" text span is defined by the 200px wide float before shape-outside image has been loaded and by the 100px wide image segment after it has been loaded.

Hello World

FAIL 'shape-outside' layout is updated after the image has been loaded assert_equals: image loaded expected 100 but got 101
